--- ChatBox/FriendsScene/Model/FriendsModel.swift
@@ -1,14 +1,0 @@
-//
-//  FriendsModel.swift
-//  ChatBox
-//
-//  Created by Alexander Chervoncev on 21/2/2023.
-//
-
-import Foundation
-
-struct FriendsModel {
-    var id: String
-    var nickname: String
-    var avatarURL: String?
-}
