+++ ChatBox/FriendsScene/View/FriendsTableViewCell.xib
@@ -13,15 +13,15 @@
         <tableViewCell contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" rowHeight="80" id="KGk-i7-Jjw" customClass="FriendsTableViewCell" customModule="ChatBox" customModuleProvider="target">
             <rect key="frame" x="0.0" y="0.0" width="320" height="80"/>
             <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" ambiguous="YES" tableViewCell="KGk-i7-Jjw" id="H2p-sc-9uM">
+            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="KGk-i7-Jjw" id="H2p-sc-9uM">
                 <rect key="frame" x="0.0" y="0.0" width="320" height="80"/>
                 <autoresizingMask key="autoresizingMask"/>
                 <subviews>
                     <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="GgJ-Pv-a7t">
-                        <rect key="frame" x="0.0" y="0.0" width="320" height="70"/>
+                        <rect key="frame" x="0.0" y="0.0" width="320" height="80"/>
                         <subviews>
                             <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFill" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="person" catalog="system" translatesAutoresizingMaskIntoConstraints="NO" id="yze-pX-CKH">
-                                <rect key="frame" x="20" y="11.666666666666668" width="50" height="47.333333333333329"/>
+                                <rect key="frame" x="20" y="16.666666666666657" width="50" height="47.333333333333329"/>
                                 <color key="backgroundColor" white="0.66666666669999997" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                                 <color key="tintColor" white="0.33333333329999998" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                                 <constraints>
@@ -29,22 +29,25 @@
                                     <constraint firstAttribute="width" constant="50" id="ipF-6Q-jrT"/>
                                 </constraints>
                             </imageView>
-                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="firstname lastname" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="UWb-E9-9Hf">
-                                <rect key="frame" x="90" y="24.666666666666671" width="220" height="21"/>
-                                <fontDescription key="fontDescription" type="system" pointSize="17"/>
+                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="firstname lastname" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="UWb-E9-9Hf">
+                                <rect key="frame" x="78" y="14" width="234" height="21"/>
+                                <fontDescription key="fontDescription" type="system" pointSize="18"/>
                                 <nil key="textColor"/>
                                 <nil key="highlightedColor"/>
                             </label>
                         </subviews>
                         <color key="backgroundColor" systemColor="systemBackgroundColor"/>
                         <constraints>
-                            <constraint firstItem="UWb-E9-9Hf" firstAttribute="leading" secondItem="yze-pX-CKH" secondAttribute="trailing" constant="20" id="4M7-wE-FPU"/>
-                            <constraint firstAttribute="height" constant="70" id="9WJ-fH-0KL"/>
-                            <constraint firstItem="yze-pX-CKH" firstAttribute="leading" secondItem="GgJ-Pv-a7t" secondAttribute="leading" constant="20" symbolic="YES" id="AZX-Xl-lZ2"/>
+                            <constraint firstAttribute="height" constant="80" id="9WJ-fH-0KL"/>
+                            <constraint firstItem="yze-pX-CKH" firstAttribute="top" secondItem="GgJ-Pv-a7t" secondAttribute="top" constant="15" id="ARB-8q-79G"/>
+                            <constraint firstItem="UWb-E9-9Hf" firstAttribute="leading" secondItem="yze-pX-CKH" secondAttribute="trailing" constant="8" id="BLA-Hj-GFU"/>
                             <constraint firstAttribute="width" constant="320" id="JAF-Gq-jOC"/>
-                            <constraint firstAttribute="trailing" secondItem="UWb-E9-9Hf" secondAttribute="trailing" constant="10" id="dAH-Yx-KZk"/>
-                            <constraint firstItem="yze-pX-CKH" firstAttribute="centerY" secondItem="GgJ-Pv-a7t" secondAttribute="centerY" id="kV3-Nk-E6M"/>
-                            <constraint firstItem="UWb-E9-9Hf" firstAttribute="centerY" secondItem="GgJ-Pv-a7t" secondAttribute="centerY" id="n9J-41-1gP"/>
+                            <constraint firstItem="UWb-E9-9Hf" firstAttribute="top" secondItem="GgJ-Pv-a7t" secondAttribute="top" constant="14" id="KyK-iP-c7n"/>
+                            <constraint firstItem="yze-pX-CKH" firstAttribute="centerY" secondItem="GgJ-Pv-a7t" secondAttribute="centerY" id="WI6-aM-QDQ"/>
+                            <constraint firstAttribute="bottom" secondItem="yze-pX-CKH" secondAttribute="bottom" constant="15" id="dB7-LE-UuZ"/>
+                            <constraint firstItem="yze-pX-CKH" firstAttribute="leading" secondItem="GgJ-Pv-a7t" secondAttribute="leading" constant="20" id="eTC-Rc-fxJ"/>
+                            <constraint firstAttribute="bottom" secondItem="UWb-E9-9Hf" secondAttribute="bottom" constant="45" id="f8l-YB-YSj"/>
+                            <constraint firstAttribute="trailing" secondItem="UWb-E9-9Hf" secondAttribute="trailing" constant="8" id="qg7-C0-RwJ"/>
                         </constraints>
                     </view>
                 </subviews>
