--- ChatBox/UserScene/Model/UserModel.swift
@@ -1,14 +1,0 @@
-//
-//  UserModel.swift
-//  ChatBox
-//
-//  Created by Alexander Chervoncev on 28/3/2023.
-//
-
-import Foundation
-
-struct UserModel {
-    var id: String
-    var nickname: String
-    var email: String
-}
